+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -5,14 +5,12 @@ import static kitchenpos.common.fixtures.MenuFixtures.MENU1_PRICE;
 import static kitchenpos.common.fixtures.MenuGroupFixtures.MENU_GROUP1_NAME;
 import static kitchenpos.common.fixtures.OrderFixtures.ORDER1_CHANGE_STATUS_REQUEST;
 import static kitchenpos.common.fixtures.OrderFixtures.ORDER1_CREATE_REQUEST;
-import static kitchenpos.common.fixtures.OrderFixtures.ORDER1_ORDER_TABLE_ID;
 import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_NUMBER_OF_GUESTS;
 import static kitchenpos.common.fixtures.TableGroupFixtures.TABLE_GROUP1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import java.util.List;
 import kitchenpos.common.ServiceTest;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.Order;
@@ -23,7 +21,6 @@ import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.dto.order.OrderChangeStatusRequest;
 import kitchenpos.dto.order.OrderCreateRequest;
-import kitchenpos.dto.order.OrderLineItemRequest;
 import kitchenpos.dto.order.OrderLineItemsResponse;
 import kitchenpos.dto.order.OrderResponse;
 import kitchenpos.dto.order.OrdersResponse;
@@ -87,26 +84,6 @@ class OrderServiceTest extends ServiceTest {
         }
 
         @Test
-        @DisplayName("요청한 주문 항목 목록의 개수와 메뉴 아이디로 조회한 메뉴의 개수가 다르면 예외가 발생한다.")
-        void throws_notSameOrderLineItemsSize() {
-            // given
-            final OrderCreateRequest orderCreateRequest = new OrderCreateRequest(ORDER1_ORDER_TABLE_ID,
-                    List.of(new OrderLineItemRequest(1L, 1L),
-                            new OrderLineItemRequest(2L, 1L)
-                    )
-            );
-            final MenuGroup menuGroup = new MenuGroup(MENU_GROUP1_NAME);
-            final MenuGroup savedMenuGroup = menuGroupRepository.save(menuGroup);
-            menuRepository.save(new Menu(MENU1_NAME, MENU1_PRICE, savedMenuGroup));
-            orderTableRepository.save(new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, false));
-
-            // when & then
-            assertThatThrownBy(() -> orderService.create(orderCreateRequest))
-                    .isInstanceOf(OrderException.NotFoundOrderLineItemMenuExistException.class)
-                    .hasMessage("[ERROR] 주문 항목 목록에 메뉴가 누락된 주문 항목이 존재합니다.");
-        }
-
-        @Test
         @DisplayName("주문 테이블 ID에 해당하는 주문 테이블이 존재하지 않으면 예외가 발생한다.")
         void throws_notFoundOrderTable() {
             // given
@@ -119,22 +96,6 @@ class OrderServiceTest extends ServiceTest {
             assertThatThrownBy(() -> orderService.create(orderCreateRequest))
                     .isInstanceOf(OrderTableException.NotFoundOrderTableException.class)
                     .hasMessage("[ERROR] 해당하는 OrderTable이 존재하지 않습니다.");
-        }
-
-        @Test
-        @DisplayName("주문 테이블이 비어있는 상태면 예외가 발생한다")
-        void throws_OrderTableIsEmpty() {
-            // given
-            final OrderCreateRequest orderCreateRequest = ORDER1_CREATE_REQUEST();
-            final MenuGroup menuGroup = new MenuGroup(MENU_GROUP1_NAME);
-            final MenuGroup savedMenuGroup = menuGroupRepository.save(menuGroup);
-            menuRepository.save(new Menu(MENU1_NAME, MENU1_PRICE, savedMenuGroup));
-            orderTableRepository.save(new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, true));
-
-            // when & then
-            assertThatThrownBy(() -> orderService.create(orderCreateRequest))
-                    .isInstanceOf(OrderException.CannotOrderStateByOrderTableEmptyException.class)
-                    .hasMessage("[ERROR] 주문 테이블이 비어있는 상태일 때 주문할 수 없습니다.");
         }
     }
